Add CCNotFoundException

+++ src/main/java/seedu/address/model/person/UniqueClassList.java
@@ -9,6 +9,7 @@ import java.util.List;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+import seedu.address.model.person.exceptions.CCNotFoundException;
 import seedu.address.model.person.exceptions.DuplicateClassException;
 
 /**
@@ -60,9 +61,9 @@ public class UniqueClassList implements Iterable<Classes> {
         requireAllNonNull(target, editedClass);
 
         int index = internalList.indexOf(target);
-//        if (index == -1) {
-//            throw new CcNotFoundException();
-//        }
+        if (index == -1) {
+            throw new CCNotFoundException();
+        }
         if (!target.isSameClass(editedClass) && contains(editedClass)) {
             throw new DuplicateClassException();
         }
@@ -89,7 +90,9 @@ public class UniqueClassList implements Iterable<Classes> {
      */
     public void remove(Classes toRemove) {
         requireNonNull(toRemove);
-        internalList.remove(toRemove);
+        if (!internalList.remove(toRemove)) {
+            throw new CCNotFoundException();
+        }
         File f = new File(toRemove.getFilePath().toUri());
         f.delete();
     }
